backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.9.1
aniso8601==6.0.0
async-timeout==4.0.2
attrs==23.1.0
autopep8==1.5.7
certifi==2022.9.24
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.8
ecdsa==0.18.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
fsspec==2023.6.0
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.205
langchainplus-sdk==0.0.11
llama-index==0.6.28
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.2
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycodestyle==2.7.0
pydantic==1.10.9
pypandoc==1.11
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rsa==4.9
six==1.12.0
SQLAlchemy==2.0.16
tenacity==8.2.2
tiktoken==0.4.0
toml==0.10.2
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.2.2
yarl==1.9.2
zipp==3.11.0
